requirements.txt
python-dotenv==0.21.0
python-telegram-bot==13.14
requests==2.31.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.6